+++ CHANGELOG.md
@@ -9,7 +9,13 @@ The format is based on [Keep a Changelog], and this project adheres to
 [semantic versioning]: https://semver.org/spec/v2.0.0.html
 [conventional commits]: https://www.conventionalcommits.org/en/v1.0.0/
 
-## [0.2.1] - 2024-04-30
+## [1.0.0] - 2024-04-30
+### Breaking changes
+ - Add configurable hook timeout
+
+### 🚀 Features
+
+- [**breaking**] Add configurable hook timeout
 
 ### ⚙️ Miscellaneous Tasks